system: core: libsparse: Fix incompatible pointer type build issue
After Upgrade GCC14 version, it report -Wincompatible-pointer-types
errors, change size_t to int to resolve this function output_file_open_callback incompatible pointer issue.

+++ libsparse/sparse.c
@@ -53,7 +53,7 @@ struct chunk_data {
   int (*write)(void* priv, const void* data, size_t len, unsigned int block, unsigned int nr_blocks);
 };
 
-static int foreach_chunk_write(void* priv, const void* data, size_t len) {
+static int foreach_chunk_write(void* priv, const void* data, int len) {
   struct chunk_data* chk = (struct chunk_data*)priv;
 
   return chk->write(chk->priv, data, len, chk->block, chk->nr_blocks);
